qjazz-server/python/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "qjazz-server"
description = "Cloud ready Qgis server services"
requires-python = ">= 3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX",
]
dependencies = [
    "qjazz-contrib==0.1.0.dev0",
    "aiohttp",
    "grpcio",
    "grpcio-tools",
    "grpcio-health-checking",
    "click",
    "dnspython",
    "jsondiff",
    "ruamel.yaml",
]
version = "0.1.0.dev0"

[project.optional-dependencies]
amqp = ["py-amqp-client>=2.0.0"]
minio = ['minio']

[[project.authors]]
name = "David Marteau"
email = "dmarteau@3liz.org"

[project.scripts]
qjazz-server-admin = "qjazz_admin.cli:main"
qjazz-config = "qjazz_rpc.cli:main"


[tool.setuptools.packages.find]
exclude = [ "tests/*" ]
include = [
    "qjazz_rpc",
    "qjazz_rpc.*",
    "qjazz_map",
    "qjazz_map.*",
    "qjazz_admin",
    "qjazz_admin.*",
    "qjazz_ogc",
    "qjazz_ogc.*",
]


[tool.autopep8]
exclude = "*_pb2.py, *_pb2_grpc.py"

[tool.ruff]
extend = "../../config/ruff.toml"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T201"]
"qjazz_http/config.py" = ["T201"]
